// packages/prettier-plugin-java/src/tree-sitter-java.d.ts
interface SyntaxNodeBase {
  tree: Tree;
  id: number;
  typeId: number;
  grammarId: number;
  type: string;
  grammarType: string;
  isNamed: boolean;
  isMissing: boolean;
  isExtra: boolean;
  hasChanges: boolean;
  hasError: boolean;
  isError: boolean;
  text: string;
  parseState: number;
  nextParseState: number;
  startPosition: Point;
  endPosition: Point;
  startIndex: number;
  endIndex: number;
  parent: SyntaxNode | null;
  children: Array<SyntaxNode>;
  namedChildren: Array<SyntaxNode>;
  childCount: number;
  namedChildCount: number;
  firstChild: SyntaxNode | null;
  firstNamedChild: SyntaxNode | null;
  lastChild: SyntaxNode | null;
  lastNamedChild: SyntaxNode | null;
  nextSibling: SyntaxNode | null;
  nextNamedSibling: SyntaxNode | null;
  previousSibling: SyntaxNode | null;
  previousNamedSibling: SyntaxNode | null;
  descendantCount: number;

  toString(): string;
  child(index: number): SyntaxNode | null;
  namedChild(index: number): SyntaxNode | null;
  childForFieldName(fieldName: string): SyntaxNode | null;
  childForFieldId(fieldId: number): SyntaxNode | null;
  fieldNameForChild(childIndex: number): string | null;
  childrenForFieldName(fieldName: string): Array<SyntaxNode>;
  childrenForFieldId(fieldId: number): Array<SyntaxNode>;
  firstChildForIndex(index: number): SyntaxNode | null;
  firstNamedChildForIndex(index: number): SyntaxNode | null;

  descendantForIndex(index: number): SyntaxNode;
  descendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  namedDescendantForIndex(index: number): SyntaxNode;
  namedDescendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  descendantForPosition(position: Point): SyntaxNode;
  descendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  namedDescendantForPosition(position: Point): SyntaxNode;
  namedDescendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  descendantsOfType<T extends TypeString>(types: T | readonly T[], startPosition?: Point, endPosition?: Point): NodeOfType<T>[];

  closest<T extends SyntaxType>(types: T | readonly T[]): NamedNode<T> | null;
  walk(): TreeCursor;
}
// ...
interface NamedNodeBase extends SyntaxNodeBase {
    isNamed: true;
}
// ...
export interface ExportsModuleDirectiveNode extends NamedNodeBase {
  type: SyntaxType.ExportsModuleDirective;
  modulesNodes: (IdentifierNode | ScopedIdentifierNode)[];
  packageNode: IdentifierNode | ScopedIdentifierNode;
}
// ...
export interface OpensModuleDirectiveNode extends NamedNodeBase {
  type: SyntaxType.OpensModuleDirective;
  modulesNodes: (IdentifierNode | ScopedIdentifierNode)[];
  packageNode: IdentifierNode | ScopedIdentifierNode;
}
// ...
export interface ScopedIdentifierNode extends NamedNodeBase {
  type: SyntaxType.ScopedIdentifier;
  nameNode: IdentifierNode;
  scopeNode: IdentifierNode | ScopedIdentifierNode;
}
// ...
export interface IdentifierNode extends NamedNodeBase {
  type: SyntaxType.Identifier;
}